requirements.txt
pyserial~=3.5

numpy~=1.26.3
pandas~=2.2.0

pynmeagps~=1.0.33

tabulate~=0.9.0
pyubx2~=1.2.36